# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TravelOS – Multi-tenant Travel ERP SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List drivers
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [drivers]
      summary: Create driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /v1/drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
    patch:
      operationId: updateDriver
      tags: [drivers]
      summary: Update driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
    delete:
      operationId: deleteDriver
      tags: [drivers]
      summary: Delete driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /v1/finance/invoices:
    get:
      operationId: listInvoices
      tags: [finance]
      summary: List invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [finance]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /v1/finance/invoices/{id}:
    get:
      operationId: getInvoice
      tags: [finance]
      summary: Get invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [finance]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /v1/finance/expenses:
    get:
      operationId: listExpenses
      tags: [finance]
      summary: List expenses
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [finance]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /v1/finance/summary:
    get:
      operationId: getFinanceSummary
      tags: [finance]
      summary: Finance P&L summary
      parameters:
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Finance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceSummary"

  # ── PUBLIC WEBSITE ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Driver:
      type: object
      required: [id, name, phone, status, licenseNumber]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, on_trip, off_duty, inactive]
        licenseNumber:
          type: string
        licenseExpiry:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        totalTrips:
          type: integer
        photoUrl:
          type: ["string", "null"]
        vehicleAssigned:
          type: ["string", "null"]
        joiningDate:
          type: ["string", "null"]
        createdAt:
          type: string
    DriverInput:
      type: object
      required: [name, phone, licenseNumber]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        licenseNumber:
          type: string
        licenseExpiry:
          type: string
        joiningDate:
          type: string
        vehicleId:
          type: string
    DriverUpdate:
      type: object
      properties:
        status:
          type: string
        licenseExpiry:
          type: string
        vehicleId:
          type: string
...
    Invoice:
      type: object
      required: [id, invoiceNumber, customerName, amount, status, dueDate]
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        customerName:
          type: string
        bookingId:
          type: ["string", "null"]
        amount:
          type: number
        taxAmount:
          type: number
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        dueDate:
          type: string
        paidAt:
          type: ["string", "null"]
        paymentMode:
          type: ["string", "null"]
        createdAt:
          type: string
    InvoiceInput:
      type: object
      required: [customerName, amount, dueDate]
      properties:
        customerName:
          type: string
        bookingId:
          type: string
        amount:
          type: number
        taxAmount:
          type: number
        dueDate:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        status:
          type: string
        paymentMode:
          type: string
        paidAt:
          type: string
    Expense:
      type: object
      required: [id, category, amount, date, description]
      properties:
        id:
          type: string
        category:
          type: string
          enum: [fuel, maintenance, salary, toll, insurance, misc]
        amount:
          type: number
        date:
          type: string
        description:
          type: string
        vendorName:
          type: ["string", "null"]
        receiptUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [category, amount, date, description]
      properties:
        category:
          type: string
        amount:
          type: number
        date:
          type: string
        description:
          type: string
        vendorName:
          type: string
    FinanceSummary:
      type: object
      required: [totalRevenue, totalExpenses, netProfit, pendingInvoices, overdueAmount]
      properties:
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        pendingInvoices:
          type: integer
        overdueAmount:
          type: number
        paidInvoices:
          type: integer
        grossMargin:
          type: number

    # ── Public Enquiry ─────────────────────────────────────────────────────
...